Move SOMD1 compatibility functions to utilities module.
