+++ src/somd2/_utils/_somd1.py
